transaction: move enrollment test from dev into the master
